agents/requirements.txt
openai==1.54.5
anthropic==0.40.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.4
pathlib==1.0.1
websockets==12.0